+++ package.json
@@ -6,7 +6,7 @@
     "scripts": {
         "start": "nodemon --exec babel-node src/app.js",
         "test": "echo \"Error: no test specified\" && exit 1"
-    }, 
+    },
     "author": "jiwon-yang",
     "license": "ISC",
     "dependencies": {
@@ -33,6 +33,7 @@
         "nodemailer": "^6.4.2",
         "passport": "^0.4.0",
         "passport-local": "^1.0.0",
+        "pug": "^2.0.4",
         "qs": "^6.9.1",
         "querystring": "^0.2.0",
         "sequelize": "^5.21.2",
