+++ src/controllers/mainController.js
@@ -4,7 +4,7 @@ const path = require('path')
 const getMain = async (req, res) => {
   try {
     const main = fs.readFileSync(
-      path.resolve(__dirname, '../public/html/index.html'),
+      path.resolve(__dirname, '../public/templates/index.html'),
       'utf8'
     )
     res.send(main)
