--- src/public/html/confirmSchool.html
@@ -1,4 +1,0 @@
-<div></div>
-<script>
-alert('인증이 완료되었습니다! 앱에서 가입을 계속 진행해주세요!')
-</script>
