--- src/public/html/findIdMail.html
@@ -1,12 +1,0 @@
-<div>
-  안녕하세요<br />
-  팅팅 아이디 찾기를 위한 이메일 입니다. <br />
-  회원님의 아이디는 다음과 같습니다. <br />
-  <br />
-  <strong></strong><br />
-  <br />
-  <br />
-  이 링크는 30분 동안 유효합니다. <br />
-  30분이 지난 후에는 다시 인증받으세요. <br /><br />
-  본 메일은 발신 전용 메일로 회신되지 않습니다.<br />
-</div>
