--- src/public/html/resetPasswordMail.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<body>
-  <table align=center style="width: 600px;text-align: center;">
-    <tr>
-      <td>
-        <img src="https://tingting-logo.s3.ap-northeast-2.amazonaws.com/web/mail/tingtinglogo.png" alt="TingTing " width="120" height="40"><br />
-        <hr style="width: 600px;color: #FF7D70;border: thin solid;">
-        <div>
-          <br />
-          안녕하세요.<br />
-          팅팅 비밀번호 재설정을 위한 <span style="color: #FF7D70;">인증 메일</span>입니다. <br /><br />
-          <strong>아래 버튼을 클릭하여 이메일 인증을 완료하세요.</strong><br /><br />
-          <hr style="width: 600px;color: #FF7D70;border: thin solid;"><br />
-          <a href="http://api.tingting.kr/api/v1/auth/find/password/confirm?code="> 
-            <button
-              style=" width: 300px;height: 35px;border: #FF7D70 2px solid;background-color: white;color: #FF7D70;text-align: center;border-radius: 4px;cursor: pointer;"
-              onMouseOver="this.style.color='#FFFFFF';this.style.backgroundColor ='#FF7D70';" onMouseOut="this.style.color='#FF7D70';this.style.backgroundColor ='#FFFFFF';">
-              인증하기
-            </button>
-          </a>
-          <p id="after"></p>
-          <p style="font-size: 10px;">* 이 링크는 30분 동안 유효합니다. 30분이 지난 후에는 다시 인증받으세요.</p><br /><br />
-          <p style="font-size: 10px;color: darkgray;">본 메일은 발신 전용 메일로 회신되지 않습니다.</p><br /><br />
-        </div>
-      </td>
-    </tr>
-  </table>
-</body>
-
-</html>
