+++ src/public/templates/authMail.pug
@@ -1,0 +1,31 @@
+doctype html
+table(align='center' style='width: 600px;text-align: center;')
+  tr
+    td
+      img(src='https://tingting-logo.s3.ap-northeast-2.amazonaws.com/web/mail/tingtinglogo.png' alt='TingTing ' width='120' height='40')
+      br
+      hr(style='width: 600px;color: #FF7D70;border: thin solid;')
+      div
+        br
+        | 안녕하세요.
+        br
+        | 팅팅 회원가입을 위한
+        span(style='color: #FF7D70;') 인증 메일
+        | 입니다.
+        br
+        br
+        strong 아래 버튼을 클릭하여 이메일 인증을 완료하세요.
+        br
+        br
+        hr(style='width: 600px;color: #FF7D70;border: thin solid;')
+        br
+        a(href=`${ authenticationUrl }`)
+          button(style=' width: 300px;height: 35px;border: #FF7D70 2px solid;background-color: white;color: #FF7D70;text-align: center;border-radius: 4px;cursor: pointer;' onmouseover="this.style.color='#FFFFFF';this.style.backgroundColor ='#FF7D70';" onmouseout="this.style.color='#FF7D70';this.style.backgroundColor ='#FFFFFF';")
+            | 인증하기
+        p#after
+        p(style='font-size: 10px;') * 이 링크는 30분 동안 유효합니다. 30분이 지난 후에는 다시 인증받으세요.
+        br
+        br
+        p(style='font-size: 10px;color: darkgray;') 본 메일은 발신 전용 메일로 회신되지 않습니다.
+        br
+        br
