+++ src/public/templates/findIdMail.pug
@@ -1,0 +1,24 @@
+doctype html
+table(align='center' style='width: 600px;text-align: center;')
+  tr
+    td
+      img(src='https://tingting-logo.s3.ap-northeast-2.amazonaws.com/web/mail/tingtinglogo.png' alt='TingTing ' width='120' height='40')
+      br
+      hr(style='width: 600px;color: #FF7D70;border: thin solid;')
+      div
+        br
+        | 안녕하세요.
+        br
+        | 팅팅 아이디 찾기를 위한 메일입니다.
+        br
+        br
+        strong 회원님의 아이디는 다음과 같습니다.
+        br
+        br
+        h2(style='font-size: 19px;') #{ localId }
+        br
+        hr(style='width: 600px;color: #FF7D70;border: thin solid;')
+        br
+        p(style='font-size: 10px;color: darkgray;') 본 메일은 발신 전용 메일로 회신되지 않습니다.
+        br
+        br
